# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/role/page": {
      "get": {
        "operationId": "RoleController_pageRole",
        "parameters": [
          {
            "name": "roleName",
            "required": true,
            "in": "query",
            "description": "角色名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "required": true,
            "in": "query",
            "description": "角色编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "query",
            "description": "角色等级",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/RolePageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询角色列表",
        "tags": ["Role"]
      }
    },
    "/role/schema-generator-holder-page-role": {
      "post": {
        "deprecated": true,
        "operationId": "RoleController__pageRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["Role"]
      }
    },
    "/role/option": {
      "get": {
        "operationId": "RoleController_pageOptionRole",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "用户ID",
            "schema": {
              "example": "",
              "type": "number"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "返回的列表字段",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/RolePageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询角色列表(可选字段)",
        "tags": ["Role"]
      }
    },
    "/role/schema-generator-holder-page-option-role": {
      "post": {
        "deprecated": true,
        "operationId": "RoleController__pageOptionRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePageOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["Role"]
      }
    },
    "/role/find/{id}": {
      "get": {
        "operationId": "RoleController_findRoleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询角色",
        "tags": ["Role"]
      }
    },
    "/role/batch/query": {
      "get": {
        "operationId": "RoleController_findRoleListByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "用“,”拼接的id",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleListRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按ids查询角色",
        "tags": ["Role"]
      }
    },
    "/role/create": {
      "post": {
        "operationId": "RoleController_createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "创建角色",
        "tags": ["Role"]
      }
    },
    "/role/update/{id}": {
      "post": {
        "operationId": "RoleController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RoleRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改角色",
        "tags": ["Role"]
      }
    },
    "/role/delete/{id}": {
      "post": {
        "operationId": "RoleController_removeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "删除角色",
        "tags": ["Role"]
      }
    },
    "/role/batch/delete": {
      "post": {
        "operationId": "RoleController_batchRemoveRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量删除角色",
        "tags": ["Role"]
      }
    },
    "/role/updateStatus/{id}": {
      "post": {
        "operationId": "RoleController_updateRoleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改角色状态",
        "tags": ["Role"]
      }
    },
    "/role/batch/status": {
      "post": {
        "operationId": "RoleController_batchUpdateRoleStatus",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchUpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量修改角色状态",
        "tags": ["Role"]
      }
    },
...
    "/role/export": {
      "post": {
        "operationId": "RoleController_exportRole",
        "parameters": [
          {
            "name": "roleName",
            "required": true,
            "in": "query",
            "description": "角色名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "required": true,
            "in": "query",
            "description": "角色编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "query",
            "description": "角色等级",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "导出角色数据",
        "tags": ["Role"]
      }
    },
...
  },
  "info": {
    "title": "BACKED API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "head": {
            "$ref": "#/components/schemas/ResponseHeadDto"
          }
        },
        "required": ["head"]
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {}
      },
...
      "BatchRespDto": {
        "type": "object",
        "properties": {
          "notFoundIds": {
            "description": "未找到的id数组",
            "example": "",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BatchDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "string",
            "description": "用“,”拼接的id",
            "example": ""
          }
        },
        "required": ["ids"]
      },
      "UpdateStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "状态值",
            "enum": [1, 0]
          }
        },
        "required": ["status"]
      },
...
      "RolePageRespDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "角色id"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "roleCode": {
            "type": "string",
            "description": "角色编码"
          },
          "level": {
            "type": "number",
            "description": "角色等级"
          },
          "status": {
            "type": "number",
            "description": "状态：0-禁用，1-启用"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间"
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间"
          },
          "userCount": {
            "type": "number",
            "description": "用户数量"
          },
          "resourceCount": {
            "type": "number",
            "description": "资源数量"
          }
        },
        "required": ["id", "roleName", "roleCode", "level", "userCount", "resourceCount"]
      },
      "RolePageDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "roleCode": {
            "type": "string",
            "description": "角色编码"
          },
          "level": {
            "type": "number",
            "description": "角色等级"
          },
          "status": {
            "type": "number",
            "description": "状态：0-禁用，1-启用"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间"
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间"
          },
          "page": {
            "type": "number",
            "description": "页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数"
          },
          "createdAtStart": {
            "type": "string",
            "description": "创建时间开始",
            "example": "2000-10-10 00:00:00"
          },
          "createdAtEnd": {
            "type": "string",
            "description": "创建时间结束",
            "example": "2000-10-10 23:59:59"
          },
          "updatedAtStart": {
            "type": "string",
            "description": "修改时间开始",
            "example": "2000-10-10 00:00:00"
          },
          "updatedAtEnd": {
            "type": "string",
            "description": "修改时间结束",
            "example": "2000-10-10 23:59:59"
          }
        },
        "required": ["roleName", "roleCode", "level"]
      },
      "RolePageOptionDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数"
          },
          "keyword": {
            "type": "string",
            "description": "关键字",
            "example": ""
          },
          "userId": {
            "type": "number",
            "description": "用户ID",
            "example": ""
          },
          "fields": {
            "type": "string",
            "description": "返回的列表字段",
            "example": ""
          }
        }
      },
      "RoleRespDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "角色id"
          },
          "roleName": {
            "type": "string",
            "description": "角色名称"
          },
          "roleCode": {
            "type": "string",
            "description": "角色编码"
          },
          "level": {
            "type": "number",
            "description": "角色等级"
          },
          "status": {
            "type": "number",
            "description": "状态：0-禁用，1-启用"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间"
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间"
          },
          "userIds": {
            "type": "string",
            "description": "用户id列表"
          },
          "resourceIds": {
            "type": "string",
            "description": "资源id列表"
          }
        },
        "required": ["id", "roleName", "roleCode", "level", "userIds", "resourceIds"]
      },
      "RoleListRespDto": {
        "type": "object",
        "properties": {
          "notFoundIds": {
            "description": "未找到的id数组",
            "example": "",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "list": {
            "description": "列表",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleRespDto"
            }
          }
        },
        "required": ["list"]
      },
...
    }
  }
}